% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{onezero}
\alias{onezero}
\title{1/0 Flags}
\usage{
onezero(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
an integer vector
}
\description{
Turns Yes/No variables into 1/0 integers
}
\details{
The input is forced to lowercase and only the first character, the "y" or
"n", is used to map to the 1/0 integer values.  The function allows for "1"
and "0" to be in the character vector as well.
}
\examples{

flag <- c("Y", "No", "NO", "no", "n", "YES", "Yes", "yEs", "1", "0")
onezero(flag)

}
