% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-pedalfast.R
\docType{data}
\name{pedalfast}
\alias{pedalfast}
\title{PEDALFAST Data}
\format{
An object of class \code{data.frame} with 388 rows and 103 columns.
}
\usage{
pedalfast
}
\description{
Single data frame for the PEDALFAST data.
}
\details{
\code{pedalfast} is one data frame with the whole of the exported data.
}
\seealso{
\code{vignette("datasets", package = "pedalfast.data")}
}
\keyword{datasets}
