% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-pedalfast.R
\docType{data}
\name{pedalfast_metadata}
\alias{pedalfast_metadata}
\title{PEDALFAST Metadata}
\format{
An object of class \code{data.frame} with 103 rows and 3 columns.
}
\usage{
pedalfast_metadata
}
\description{
\code{pedalfast_metadata} provides documentation for each of variables in the
\code{pedalfast}.
}
\seealso{
\code{vignette("datasets", package = "pedalfast.data")}
}
\keyword{datasets}
