% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{yesno}
\alias{yesno}
\title{Yes No Flags}
\usage{
yesno(x)
}
\arguments{
\item{x}{an integer vector}
}
\value{
a character vector
}
\description{
Turns 1/0 into "Yes"/"No"
}
\examples{

flag <- c(0, 1, 1, 0, 0, 0)
yesno(flag)

}
